Move Store model definition to AbstractStore
