+++ pootle/apps/pootle_store/abstracts.py
@@ -9,12 +9,24 @@
 from translate.storage import base
 
 from django.conf import settings
+from django.core.validators import MinValueValidator
 from django.db import models
 
+from pootle.core.mixins import CachedTreeItem
+from pootle.core.storage import PootleFileSystemStorage
 from pootle.core.user import get_system_user, get_system_user_id
-from .constants import UNTRANSLATED
+from pootle.core.utils.timezone import datetime_min
+from pootle_format.models import Format
+from pootle.i18n.gettext import ugettext_lazy as _
 
-from .fields import MultiStringField
+from .constants import NEW, UNTRANSLATED
+from .fields import MultiStringField, TranslationStoreField
+from .managers import StoreManager
+from .validators import validate_no_slashes
+
+
+# Needed to alter storage location in tests
+fs = PootleFileSystemStorage()
 
 
 class AbstractUnit(models.Model, base.TranslationUnit):
@@ -92,4 +104,97 @@ class AbstractUnit(models.Model, base.TranslationUnit):
     reviewed_on = models.DateTimeField(db_index=True, null=True)
 
     class Meta(object):
+        abstract = True
+
+
+class AbstractStore(models.Model, CachedTreeItem, base.TranslationStore):
+
+    file = TranslationStoreField(
+        max_length=255,
+        storage=fs,
+        db_index=True,
+        null=False,
+        editable=False)
+
+    parent = models.ForeignKey(
+        'pootle_app.Directory',
+        related_name='child_stores',
+        db_index=True,
+        editable=False,
+        on_delete=models.CASCADE)
+
+    translation_project_fk = 'pootle_translationproject.TranslationProject'
+    translation_project = models.ForeignKey(
+        translation_project_fk,
+        related_name='stores',
+        db_index=True,
+        editable=False,
+        on_delete=models.CASCADE)
+
+    filetype = models.ForeignKey(
+        Format,
+        related_name='stores',
+        null=True,
+        blank=True,
+        db_index=True,
+        on_delete=models.CASCADE)
+    is_template = models.BooleanField(default=False)
+
+    # any changes to the `pootle_path` field may require updating the schema
+    # see migration 0007_case_sensitive_schema.py
+    pootle_path = models.CharField(
+        max_length=255,
+        null=False,
+        unique=True,
+        db_index=True,
+        verbose_name=_("Path"))
+
+    tp_path = models.CharField(
+        max_length=255,
+        null=True,
+        blank=True,
+        db_index=True,
+        verbose_name=_("Path"))
+    # any changes to the `name` field may require updating the schema
+    # see migration 0007_case_sensitive_schema.py
+    name = models.CharField(
+        max_length=128,
+        null=False,
+        editable=False,
+        validators=[validate_no_slashes])
+
+    file_mtime = models.DateTimeField(default=datetime_min)
+    state = models.IntegerField(
+        null=False,
+        default=NEW,
+        editable=False,
+        db_index=True)
+    creation_time = models.DateTimeField(
+        auto_now_add=True,
+        db_index=True,
+        editable=False,
+        null=True)
+    last_sync_revision = models.IntegerField(
+        db_index=True,
+        null=True,
+        blank=True)
+    obsolete = models.BooleanField(default=False)
+
+    # this is calculated from virtualfolders if installed and linked
+    priority = models.FloatField(
+        db_index=True,
+        default=1,
+        validators=[MinValueValidator(0)])
+
+    objects = StoreManager()
+
+    class Meta(object):
+        ordering = ['pootle_path']
+        index_together = [
+            ["translation_project", "is_template"],
+            ["translation_project", "pootle_path", "is_template", "filetype"]]
+        unique_together = (
+            ('parent', 'name'),
+            ("obsolete", "translation_project", "tp_path"))
+        base_manager_name = "objects"
         abstract = True
